+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 ## 🚀 Features
 - `create` now has `--env` option to load in an .env file
+- made inputs and outputs prefix optional in `connect`
 ## 🐛 Bugfixes
 - default values are used on input connections